Docs command can now add multiple links (up to 10).
Also fixed a bug where the help message could render an empty MessageEmbed() field when the last field is supposed to have only one item. -> The current fix is not checking whether the last field contains more than 1024 characters, so there's this chance it may throw an error for this reason.